Move checkpoints to another file
